Use `Version#to_file` instead of reimplementing inline.
The punch CLI was writing `Version` info back to the version file
inline. This was leading to the `DateVersionPart` bug sticking around
for command-line usage, even though `Version#to_file` was already fixed.
Since there was already an instance method on `Version` for
serializing that data to a file, I made the CLI use that method,
instead of reimplementing the logic inline.